requirements.txt
Django==1.5
South
psycopg2
dj-database-url
django-mptt
